# openapi/finance-api.yaml
openapi: 3.0.3
info:
  title: Personal Finance Management API
  description: A comprehensive REST API for managing personal finances including transactions, budgets, recurring transactions, and analytics
  version: 1.0.0
  contact:
    name: API Support
    email: support@finance.app
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.finance.app/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user information
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallets:
    get:
      tags:
        - Auth
      summary: Get all wallets for current user
      responses:
        '200':
          description: List of wallets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WalletResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Auth
      summary: Create a new wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletCreateRequest'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wallets/{wallet_id}:
    get:
      tags:
        - Auth
      summary: Get wallet by ID
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Auth
      summary: Update wallet
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletUpdateRequest'
      responses:
        '200':
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Auth
      summary: Delete wallet
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Wallet deleted
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /budgets:
    get:
      tags:
        - Budgets
      summary: Get all budgets for current user
      responses:
        '200':
          description: List of budgets with usage
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetResponse'
    post:
      tags:
        - Budgets
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetCreateRequest'
      responses:
        '201':
          description: Budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /budgets/{budget_id}:
    get:
      tags:
        - Budgets
      summary: Get budget by ID
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Budget details with usage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Budgets
      summary: Update budget
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetUpdateRequest'
      responses:
        '200':
          description: Budget updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '404':
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Budgets
      summary: Delete budget
      parameters:
        - name: budget_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Budget deleted
        '404':
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /budgets/summary:
    get:
      tags:
        - Budgets
      summary: Get budget summary for a period
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: quarter
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Budget summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetSummaryResponse'
...
  /analytics/spending-by-category:
    get:
      tags:
        - Analytics
      summary: Get spending breakdown by category
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Spending by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategorySpendingResponse'
  /analytics/spending-trend:
    get:
      tags:
        - Analytics
      summary: Get spending trend over time
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: group_by
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: month
      responses:
        '200':
          description: Spending trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpendingTrendResponse'
  /analytics/top-categories:
    get:
      tags:
        - Analytics
      summary: Get top spending categories
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top spending categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopCategoryResponse'
  /reports/transactions.csv:
    get:
      tags:
        - Analytics
      summary: Export transactions as CSV
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /reports/transactions.pdf:
    get:
      tags:
        - Analytics
      summary: Export transactions as PDF
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
    UserRegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
      required:
        - email
        - password
        - full_name
    UserLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - full_name
        - created_at
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        user:
          $ref: '#/components/schemas/UserResponse'
      required:
        - access_token
        - token_type
        - user
    WalletCreateRequest:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
          default: VND
        initial_balance:
          type: number
          format: double
          default: 0
      required:
        - name
    WalletUpdateRequest:
      type: object
      properties:
        name:
          type: string
        currency:
          type: string
    WalletResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        currency:
          type: string
        balance:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - name
        - currency
        - balance
        - created_at
...
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [EXPENSE, INCOME]
        icon:
          type: string
        color:
          type: string
      required:
        - id
        - name
        - type
...
    BudgetCreateRequest:
      type: object
      properties:
        category_id:
          type: integer
        amount_limit:
          type: number
          format: double
        period_type:
          type: string
          enum: [MONTHLY, WEEKLY, CUSTOM]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          description: Required for CUSTOM period type
      required:
        - category_id
        - amount_limit
        - period_type
    BudgetUpdateRequest:
      type: object
      properties:
        amount_limit:
          type: number
          format: double
        period_type:
          type: string
          enum: [MONTHLY, WEEKLY, CUSTOM]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    BudgetResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        category_id:
          type: integer
        category:
          $ref: '#/components/schemas/CategoryResponse'
        amount_limit:
          type: number
          format: double
        period_type:
          type: string
          enum: [MONTHLY, WEEKLY, CUSTOM]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        used_amount:
          type: number
          format: double
        remaining_amount:
          type: number
          format: double
        usage_percentage:
          type: number
          format: double
        is_near_limit:
          type: boolean
        is_exceeded:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - category_id
        - amount_limit
        - period_type
        - used_amount
        - remaining_amount
        - usage_percentage
        - is_near_limit
        - is_exceeded
        - created_at
    BudgetSummaryResponse:
      type: object
      properties:
        period:
          type: string
        total_budget:
          type: number
          format: double
        total_spent:
          type: number
          format: double
        budgets:
          type: array
          items:
            $ref: '#/components/schemas/BudgetResponse'
...
    CategorySpendingResponse:
      type: object
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        total_amount:
          type: number
          format: double
        percentage:
          type: number
          format: double
        transaction_count:
          type: integer
    SpendingTrendResponse:
      type: object
      properties:
        period:
          type: string
        year:
          type: integer
        month:
          type: integer
        week:
          type: integer
        day:
          type: string
          format: date
        total_expense:
          type: number
          format: double
        total_income:
          type: number
          format: double
    TopCategoryResponse:
      type: object
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        total_amount:
          type: number
          format: double
        transaction_count:
          type: integer